服务端修改

+++ server/include/mysql_c_aip.h
@@ -3,7 +3,7 @@
 
 #include <mysql/mysql.h>
 MYSQL *mysql_conn_init();
-int mysql_conn_query(MYSQL *my,const char *command,int command_size,MYSQL_RES *ret_data);
+int mysql_conn_query(MYSQL *my,const char *command,long command_size,MYSQL_RES *ret_data);
 void mysql_conn_close(MYSQL *my);
 
 int mysql_result_one(MYSQL *my,const char *command,long command_size,char *buf);
